@@ -62,19 +62,19 @@
     "zone.js": "0.8.14"
   },
   "devDependencies": {
-    "@angular/animations": "4.3.5",
+    "@angular/animations": "4.4.3",
     "@angular/cli": "1.3.0-rc.3",
-    "@angular/common": "4.3.5",
-    "@angular/compiler": "4.3.5",
-    "@angular/compiler-cli": "4.3.5",
-    "@angular/core": "4.3.5",
-    "@angular/forms": "4.3.5",
-    "@angular/http": "4.3.5",
-    "@angular/language-service": "4.3.5",
-    "@angular/platform-browser": "4.3.5",
-    "@angular/platform-browser-dynamic": "4.3.5",
-    "@angular/platform-server": "4.3.5",
-    "@angular/router": "4.3.5",
+    "@angular/common": "4.4.3",
+    "@angular/compiler": "4.4.3",
+    "@angular/compiler-cli": "4.4.3",
+    "@angular/core": "4.4.3",
+    "@angular/forms": "4.4.3",
+    "@angular/http": "4.4.3",
+    "@angular/language-service": "4.4.3",
+    "@angular/platform-browser": "4.4.3",
+    "@angular/platform-browser-dynamic": "4.4.3",
+    "@angular/platform-server": "4.4.3",
+    "@angular/router": "4.4.3",
     "@ngrx/db": "2.0.1",
     "@ngrx/effects": "^4.0.2",
     "@ngrx/router-store": "^4.0.2",
