@@ -1,23 +1,28 @@
-import { BrowserModule } from '@angular/platform-browser'
+import { HTTP_INTERCEPTORS } from '@angular/common/http'
 import { NgModule } from '@angular/core'
+import { BrowserModule } from '@angular/platform-browser'
 import { RouterModule } from '@angular/router'
+import { SharedModule } from './shared/shared.module'
 
-import { SiteModule } from './site/site.module'
 import { ShopModule } from './shop/shop.module'
+import { SiteModule } from './site/site.module'
 
 import { AppComponent } from './app.component'
 
 @NgModule({
-  declarations: [
-    AppComponent,
-  ],
-  imports: [
-    BrowserModule.withServerTransition({appId: 'time-client-universal'}),
-    SiteModule,
-    ShopModule,
-    RouterModule.forRoot([]),
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
+    declarations: [
+        AppComponent,
+    ],
+    imports: [
+        BrowserModule,
+        // BrowserModule.withServerTransition({appId: 'time-client-universal'}),
+        RouterModule.forRoot([]),
+        SiteModule,
+        ShopModule,
+    ],
+    providers: [
+
+    ],
+    bootstrap: [AppComponent],
 })
 export class AppModule { }
