@@ -1,13 +1,13 @@
 import { NgModule } from '@angular/core'
 import { ServerModule } from '@angular/platform-server'
 
-import { AppModule } from './app.module'
-import { SiteModule } from './site/site.module'
-import { ShopModule } from './shop/shop.module'
 import { AppComponent } from './app.component'
+import { AppModule } from './app.module'
+import { ShopModule } from './shop/shop.module'
+import { SiteModule } from './site/site.module'
 
-import { HomeComponent } from './site/components/home/home.component'
 import { ShopComponent } from './shop/components/shop/shop.component'
+import { HomeComponent } from './site/components/home/home.component'
 
 @NgModule({
   imports: [
@@ -16,7 +16,7 @@ import { ShopComponent } from './shop/components/shop/shop.component'
     AppModule,
     ServerModule,
   ],
-  
+
   // Since the bootstrapped component is not inherited from your
   // imported AppModule, it needs to be repeated here.
   bootstrap: [AppComponent],
