@@ -1,5 +1,5 @@
-import { NgModule } from '@angular/core'
 import { CommonModule } from '@angular/common'
+import { NgModule } from '@angular/core'
 import { RouterModule } from '@angular/router'
 
 import { SharedModule } from '../shared/shared.module'
@@ -10,8 +10,8 @@ import { HomeComponent } from './components/home/home.component'
 @NgModule({
   imports: [
     CommonModule,
-    SiteRoutingModule,
     SharedModule.forRoot(),
+    SiteRoutingModule,
   ],
   declarations: [
     HomeComponent,
